refactor : split BackupService into Query, Process, Validation
